+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Fixes
 
 - [#1395](https://github.com/okta/okta-auth-js/pull/1395) Changes resolve value of `closeSession()` and `signOut()` to boolean
+- [#1398](https://github.com/okta/okta-auth-js/pull/1398) Fixes race condition in `LeaderElectionService` start
+- [#1404](https://github.com/okta/okta-auth-js/pull/1404) Adds `react-native` to `package.json`
 
 ## 7.2.0